try putting PID-file into a writeable directory
when run as daemon (started as root), we want to drop root privileges
asap; unfortunately this also means, that we can no longer write
PID-files into /var/run.
solution: create a PID-directory /var/run/WILMA, that is writeable by the user.

implementation: check whether the PID-directory exists, and if it doesn't,
create it, owned by the user. (this should prevent a chown on /var/run)

we also try to remove the PID-directory on exit